requirement.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.4
django-import-export==3.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.4.1
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.7.1
urllib3==2.0.4
xlrd==2.0.1
xlwt==1.3.0
